Moved policy/agent class to core subfolder
